Add files via upload
Use npm init in terminal and you will have to also use npm i qr-image and npm i inquirer in the terminal to install both of the dependencies to use the code. Then you can generate the qr image by entering the text or complete url of a website to get the .png image file.